+++ tests/fixtures/application_fixtures.py
@@ -2700,6 +2700,7 @@ def example_datasets() -> List[Dict]:
         "data/dataset/google_cloud_sql_mysql_example_test_dataset.yml",
         "data/dataset/google_cloud_sql_postgres_example_test_dataset.yml",
         "data/dataset/scylladb_example_test_dataset.yml",
+        "data/dataset/bigquery_enterprise_test_dataset.yml",
     ]
     for filename in example_filenames:
         example_datasets += load_dataset(filename)